feat(data/geo_importer): dockerisation of the entier geojson app and reorganizing the repository
