--- data/.env.example
@@ -1,15 +1,0 @@
-# Configuration PostgreSQL
-POSTGRES_DB=your_database_name
-POSTGRES_USER=your_database_user
-POSTGRES_PASSWORD=your_database_password
-POSTGRES_HOST=your_database_host
-POSTGRES_PORT=your_database_port
-
-# Configuration Spark
-SPARK_APP_NAME=YourSparkAppName
-SPARK_JARS=/path/to/your/jar/postgresql-42.7.4.jar
-
-# Chemins HDFS pour les fichiers GeoJSON
-HDFS_CITIES_PATH=hdfs://your_hdfs_path/geojson/communes/communes-1000m.geojson
-HDFS_DEPARTMENTS_PATH=hdfs://your_hdfs_path/geojson/departements/departements-1000m.geojson
-HDFS_REGIONS_PATH=hdfs://your_hdfs_path/geojson/regions/regions-1000m.geojson
