--- data/requirements.txt
@@ -1,4 +1,0 @@
-python-dotenv
-pyspark
-shapely
-psycopg2-binary
